state machine skeleton and bugfix for enemy approaching from top behavior
