+++ Assets/Scenes/scene_background.unity
@@ -368,7 +368,7 @@ Transform:
   m_GameObject: {fileID: 49216583}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -2.8771017, y: -0.9341165, z: 0}
+  m_LocalPosition: {x: -13.09, y: -3.92, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
