+++ Assets/Stan/script/EnemyMovement.cs
@@ -2,28 +2,23 @@ using UnityEngine;
 
 public class EnemyMovement : MonoBehaviour
 {
-    public float speed = 5f; // Speed of enemy movement
-    public float minimumDistance = 2f; // Minimum distance to maintain between enemies
+    public float speed = 5f;
     public Vector3 targetPoint; // Point outside the game scene where enemies will move if no players are present
 
-    private GameObject[] player1Objects; // Array to store player1 GameObjects
-    private GameObject[] player2Objects; // Array to store player2 GameObjects
+    private GameObject[] player1Objects;
+    private GameObject[] player2Objects;
 
     void Start()
     {
-        // Find all GameObjects tagged as "Player1" and "Player2" in the scene
         UpdatePlayerArrays();
     }
 
     void Update()
     {
-        // Update the player arrays if necessary
         UpdatePlayerArrays();
 
-        // Find the nearest player
         GameObject nearestPlayer = FindNearestPlayer();
 
-        // If a nearest player is found, move toward it
         if (nearestPlayer != null)
         {
             // Calculate direction towards the player
@@ -32,12 +27,8 @@ public class EnemyMovement : MonoBehaviour
             // Ignore the Z component to move only along the X and Y axes
             direction.z = 0f;
 
-            // Check if there's an obstacle in the way
-            if (!IsObstacleInWay(direction))
-            {
-                // Move the enemy towards the player
-                transform.Translate(direction * speed * Time.deltaTime);
-            }
+            // Move the enemy towards the player
+            transform.Translate(direction * speed * Time.deltaTime);
         }
         else
         {
@@ -91,21 +82,5 @@ public class EnemyMovement : MonoBehaviour
         }
 
         return nearestPlayer;
-    }
-
-    bool IsObstacleInWay(Vector3 direction)
-    {
-        RaycastHit2D hit;
-        // Cast a ray to check for obstacles in the specified direction
-        hit = Physics2D.Raycast(transform.position, direction, minimumDistance);
-        if (hit.collider != null)
-        {
-            // If an obstacle is found, return true
-            if (hit.collider.CompareTag("Enemy"))
-            {
-                return true;
-            }
-        }
-        return false;
     }
 }
