+++ Assets/Stan/script/LaserColliderP2.cs
@@ -1,5 +1,4 @@
 using UnityEngine;
-using System.Collections;
 
 public class LaserColliderP2 : MonoBehaviour
 {
