+++ Assets/Stan/script/PlayerBaseStates.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public abstract class PlayerBaseStates
+{
+    
+}
