+++ Assets/Stan/script/PlayerBaseStates.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: 2b3a328889d6c3f46b2f330d3ebee171
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
