+++ Assets/Stan/script/import.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d50cf60123c8d4d418df988071fa22bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
