+++ Assets/Stan/script/import/attackEffect.cs
