+++ Assets/Stan/script/import/battleManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: 6c667ca93eea6ce43bd8dd3f398b31de
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
