+++ Assets/Stan/script/state-machine.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41a9e8cd49eed564686e7e6d04e8d11d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
