+++ Assets/Stan/script/state-machine/PlayerActiveState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: ebef68ae76387b147b82acef6b0ce34e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
