+++ Assets/Stan/script/state-machine/PlayerBaseState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: 8cc213c3a90235740a49e0cca6501269
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
