+++ Assets/Stan/script/state-machine/PlayerLostState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: 8c4bba51c3da613458ce24f45721d44d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
