+++ Assets/Stan/script/state-machine/PlayerRespawnState.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+public class PlayerRespawnState : PlayerBaseState
+{
+    public override void EnterState(PlayerStateMachine player)
+    {
+
+    }
+    public override void UpdateState(PlayerStateMachine player)
+    {
+
+    }
+    public override void OnCollisionEnter(PlayerStateMachine player)
+    {
+
+    }
+}
