+++ Assets/Stan/script/state-machine/PlayerRespawnState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: b34a074e350effa4799b89bb5e47b2d9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
