+++ Assets/Stan/script/state-machine/PlayerStateMachine.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f56e1103a21f75d469a52a2b59435270
+guid: 3c1ffcb300475cf4fadbaccd19aab78e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
