+++ Assets/Stan/style/EvilEye.prefab
@@ -27,7 +27,7 @@ Transform:
   m_GameObject: {fileID: 2480370066648120274}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -0.13999999, y: 3.4100003, z: -1.6387887}
+  m_LocalPosition: {x: 0.019999743, y: -0.0100002885, z: -1.6387887}
   m_LocalScale: {x: 3, y: 3, z: 3}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -112,7 +112,7 @@ Transform:
   m_GameObject: {fileID: 2826108809213010886}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 3.4799995, y: -2.4699993, z: 0}
+  m_LocalPosition: {x: 2.99, y: -1.991, z: 0}
   m_LocalScale: {x: 0.4, y: 0.4, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -146,7 +146,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 2
-  minimumDistance: 0.1
   targetPoint: {x: -14, y: 20, z: 0}
 --- !u!70 &2864135247026438344
 CapsuleCollider2D:
@@ -180,8 +179,8 @@ CapsuleCollider2D:
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: -0.13512932, y: 3.2516572}
-  m_Size: {x: 3.04416, y: 2.2837214}
+  m_Offset: {x: 0.02035081, y: -0.06184206}
+  m_Size: {x: 3.138871, y: 2.186623}
   m_Direction: 1
 --- !u!50 &6020555218784713771
 Rigidbody2D:
@@ -237,7 +236,7 @@ Transform:
   m_GameObject: {fileID: 4004104936175885101}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -0.13999999, y: 3.4100003, z: -1.6387887}
+  m_LocalPosition: {x: 0.019999743, y: -0.0100002885, z: -1.6387887}
   m_LocalScale: {x: 3, y: 3, z: 3}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -320,7 +319,7 @@ Transform:
   m_GameObject: {fileID: 4530116871719104373}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: -0.13999999, y: 3.4100003, z: -1.6387887}
+  m_LocalPosition: {x: 0.02, y: -0.01, z: -1.6387887}
   m_LocalScale: {x: 3, y: 3, z: 3}
   m_ConstrainProportionsScale: 0
   m_Children: []
